+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Fixed
 
+- `allTests` does not generate reports if it is run by `build` or `check`
+
 ### Updated
 
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.37`